\name{makePolyFit}
\alias{makePolyFit}
\title{Ajuste polinomial.}
\usage{
  makePolyFit(degree)
}
\arguments{
  \item{degree}{Grau do polinômio de ajuste.}
}
\value{
  Função que retorna returna uma função de ajuste
  polinomial.
}
\description{
  Cria função que usa o método de mínimos quadrados para
  fazer um ajuste polinomial
}

